Intermediate checkin; created new Scheduler library
