+++ app/nidod.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python
 
-import sys, time, signal, os
+import sys
 from datetime import datetime
 from lib.Daemon import Daemon
 from lib.Nido import Config, Controller
+from lib.Scheduler import NidoSchedulerService
 from apscheduler.schedulers.background import BackgroundScheduler
 from apscheduler.jobstores.sqlalchemy import SQLAlchemyJobStore
-import rpyc
 from rpyc.utils.server import ThreadedServer
 
 class NidoDaemon(Daemon):
@@ -37,7 +37,7 @@ class NidoDaemon(Daemon):
 
         # Start scheduler and RPyC service
         self.scheduler.start()
-        server = ThreadedServer(SchedulerService(self.scheduler), port=49152, protocol_config={'allow_public_attrs': True})
+        server = ThreadedServer(NidoSchedulerService(self.scheduler), port=49152, protocol_config={'allow_public_attrs': True})
         server.start()
 
     def quit(self):
@@ -46,40 +46,6 @@ class NidoDaemon(Daemon):
         sys.stdout.write('{} [Info] Nido daemon shutdown\n'.format(datetime.utcnow()))
         sys.stdout.flush()
         return
-
-# Adapted from: https://github.com/agronholm/apscheduler/blob/master/examples/rpc/server.py
-#
-class SchedulerService(rpyc.Service):
-    def __init__(self, scheduler):
-        self._scheduler = scheduler
-
-    def __call__(self, conn):
-        return self.__class__(self._scheduler)
-
-    def exposed_add_job(self, func, *args, **kwargs):
-        return self._scheduler.add_job(func, *args, **kwargs)
-
-    def exposed_modify_job(self, job_id, jobstore=None, **changes):
-        return self._scheduler.modify_job(job_id, jobstore, **changes)
-
-    def exposed_reschedule_job(self, job_id, jobstore=None, trigger=None, **trigger_args):
-        return self._scheduler.reschedule_job(job_id, jobstore, trigger, **trigger_args)
-
-    def exposed_pause_job(self, job_id, jobstore=None):
-        return self._scheduler.pause_job(job_id, jobstore)
-
-    def exposed_resume_job(self, job_id, jobstore=None):
-        return self._scheduler.resume_job(job_id, jobstore)
-
-    def exposed_remove_job(self, job_id, jobstore=None):
-        self._scheduler.remove_job(job_id, jobstore)
-
-    def exposed_get_job(self, job_id):
-        return self._scheduler.get_job(job_id)
-
-    def exposed_get_jobs(self, jobstore=None):
-        return self._scheduler.get_jobs(jobstore)
-
 
 if __name__ == '__main__':
     config = Config().get_config()
